reorganize files into correct packages
