--- src/main/kotlin/console/ConsoleIO.kt
@@ -1,6 +1,0 @@
-package console
-
-interface ConsoleIO {
-    fun println(line:String)
-    fun readFromUser():String
-}
