--- src/main/kotlin/console/ConsoleIoImpl.kt
@@ -1,6 +1,0 @@
-package console
-
-class ConsoleIoImpl:ConsoleIO {
-    override fun println(line: String)  =  kotlin.io.println(line)
-    override fun readFromUser(): String = readln()
-}
