+++ src/main/kotlin/logic/usecases/project/GetProjectDetailsUseCase.kt
@@ -3,7 +3,7 @@ package logic.usecases.project
 import logic.entities.Project
 import logic.exception.PlanMateException.ValidationException.InvalidUUIDFormatException
 import logic.repository.ProjectRepository
-import logic.usecases.ValidateInputUseCase
+import logic.usecases.validation.ValidateInputUseCase
 import logic.usecases.task.GetAllTasksByProjectIdUseCase
 import java.util.UUID
