+++ src/main/kotlin/logic/usecases/task/CreateTaskUseCase.kt
@@ -8,7 +8,7 @@ import logic.exception.PlanMateException.ValidationException.InvalidProjectIDExc
 import logic.exception.PlanMateException.NotFoundException.StateNotFoundException
 import logic.repository.ProjectRepository
 import logic.repository.TaskRepository
-import logic.usecases.ValidateInputUseCase
+import logic.usecases.validation.ValidateInputUseCase
 import java.util.UUID
 
 class CreateTaskUseCase(
