+++ src/main/kotlin/logic/usecases/task/GetAllTasksUseCase.kt
@@ -1,4 +1,4 @@
-package logic.usecases
+package logic.usecases.task
 
 class GetAllTasksUseCase {
 }
