+++ src/main/kotlin/ui/controller/auth/CreateUserUIController.kt
@@ -1,8 +1,8 @@
-package ui.controllers
+package ui.controller.auth
 
 import logic.usecases.user.CreateUserUseCase
 import ui.console.ConsoleIO
-import ui.controller.BaseUIController
+import ui.controller.base.BaseUIController
 import ui.utils.tryToExecute
 
 class CreateUserUIController(
