+++ src/main/kotlin/ui/controller/base/BaseUIController.kt
@@ -1,4 +1,4 @@
-package ui.controller
+package ui.controller.base
 
 interface BaseUIController {
     fun execute()
