+++ src/main/kotlin/ui/controller/project/GetProjectUIController.kt
@@ -1,16 +1,16 @@
-package ui.project
+package ui.controller.project
 
 
 import logic.usecases.project.GetProjectDetailsUseCase
 import ui.console.ConsoleIO
-import ui.controller.BaseUIController
+import ui.controller.base.BaseUIController
 import ui.utils.tryToExecute
-import utils.printSwimlanesView
+import ui.utils.printSwimlanesView
 
 class GetProjectUIController(
     private val getProjectsUseCase: GetProjectDetailsUseCase,
     private val consoleIO: ConsoleIO
-):BaseUIController {
+): BaseUIController {
 
     override fun execute(){
         consoleIO.println("Enter Project ID :)")
