+++ src/main/kotlin/ui/controller/project/ViewAllProjectsUIController.kt
@@ -1,9 +1,9 @@
-package ui.project
+package ui.controller.project
 
 import logic.usecases.project.GetAllProjectsUseCase
-import ui.controller.BaseUIController
+import ui.controller.base.BaseUIController
 import ui.utils.tryToExecute
-import utils.printSwimlanesView
+import ui.utils.printSwimlanesView
 
 class ViewAllProjectsUIController(
     private val getAllProjectsUseCase: GetAllProjectsUseCase
