+++ src/main/kotlin/ui/utils/FormatLogItem.kt
@@ -1,4 +1,4 @@
-package utils
+package ui.utils
 
 import logic.entities.LogItem
