+++ src/main/kotlin/ui/utils/ResultStatus.kt
@@ -1,4 +1,4 @@
-package utils
+package ui.utils
 
 import logic.exception.PlanMateException
