+++ src/test/kotlin/ui/project/CreateProjectUIControllerTest.kt
@@ -6,6 +6,7 @@ import logic.usecases.project.CreateProjectUseCase
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import ui.controller.project.CreateProjectUIController
 import java.io.ByteArrayInputStream
 import java.io.ByteArrayOutputStream
 import java.io.PrintStream
