+++ src/test/kotlin/ui/project/GetProjectUIControllerTest.kt
@@ -10,8 +10,9 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
 import ui.console.ConsoleIO
+import ui.controller.project.GetProjectUIController
 import ui.utils.getErrorMessageByException
-import utils.printSwimlanesView
+import ui.utils.printSwimlanesView
 import java.util.*
 import kotlin.test.Test
