+++ src/test/kotlin/ui/state/UpdateStateUiControllerTest.kt
@@ -2,7 +2,6 @@ package ui.state
 
 import com.google.common.truth.Truth.assertThat
 import io.mockk.*
-import logic.exception.PlanMateException
 import logic.exception.PlanMateException.ValidationException.InvalidStateNameException
 import logic.exception.PlanMateException.ValidationException.InvalidProjectIDException
 import logic.exception.PlanMateException.ValidationException.SameStateNameException
@@ -12,7 +11,7 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
 import ui.console.ConsoleIO
-import ui.controllers.UpdateStateUiController
+import ui.controller.state.UpdateStateUiController
 import ui.utils.getErrorMessageByException
 
 class UpdateStateUiControllerTest {
